test(models): move the FSMv2CPUEnabled table into feature_usage_test.go
The table was the only thing in a file of its own, next to an existing
feature_usage_test.go covering the same struct. All five entries move
across unchanged.
